Tread schemaId as string (should be more compatible) Fix Dockerfile build fails

+++ src/main/kotlin/de/konqi/roborockbridge/persistence/DataAccessLayer.kt
@@ -85,12 +85,12 @@ class DataAccessLayer(
             }.map {
                 DeviceState(
                     device = newDevice,
-                    schemaId = it.id.toInt(),
+                    schemaId = it.id,
                     code = it.code,
                     mode = ProtocolMode.valueOf(it.mode.uppercase()),
                     type = it.type,
                     property = it.property,
-                    value = device.deviceStatus[it.id.toInt()] ?: -1
+                    value = device.deviceStatus[it.id] ?: -1
                 )
             })
